Split worktree dialog state and chrome
