+++ Tests/QuillCodeParityTests/ParityWorkspaceSettingsSheetGateTests.swift
@@ -8,6 +8,8 @@ final class ParityWorkspaceSettingsSheetGateTests: QuillCodeParityTestCase {
         let commandPaletteText = try Self.appSourceText(named: "QuillCodeCommandPaletteDialog.swift")
         let searchShortcutText = try Self.appSourceText(named: "QuillCodeSearchAndShortcutDialogs.swift")
         let worktreeDialogsText = try Self.appSourceText(named: "QuillCodeWorktreeDialogs.swift")
+        let worktreeDraftsText = try Self.appSourceText(named: "QuillCodeWorktreeDrafts.swift")
+        let worktreeChromeText = try Self.appSourceText(named: "QuillCodeWorktreeDialogChrome.swift")
         let dialogChromeText = try Self.appSourceText(named: "QuillCodeDialogChrome.swift")
 
         XCTAssertTrue(sheetsText.contains("struct QuillCodeWorkspaceSheetsModifier"), "Workspace sheet presentation should live in a focused modifier.")
@@ -27,6 +29,17 @@ final class ParityWorkspaceSettingsSheetGateTests: QuillCodeParityTestCase {
         XCTAssertTrue(searchShortcutText.contains("struct QuillCodeKeyboardShortcutsView"), "Keyboard shortcut dialog UI should live with shortcut/search dialogs.")
         XCTAssertTrue(worktreeDialogsText.contains("struct QuillCodeWorktreeCreateView"), "Worktree create UI should live in the worktree dialog file.")
         XCTAssertTrue(worktreeDialogsText.contains("struct QuillCodeWorktreeRemoveView"), "Worktree remove UI should live in the worktree dialog file.")
+        XCTAssertTrue(worktreeDraftsText.contains("struct QuillCodeWorktreeCreateDraft"), "Worktree draft/request state should live in a focused value file.")
+        XCTAssertTrue(worktreeDraftsText.contains("struct QuillCodeWorktreeOpenDraft"), "Worktree open draft state should live in a focused value file.")
+        XCTAssertTrue(worktreeDraftsText.contains("struct QuillCodeWorktreeRemoveDraft"), "Worktree remove draft state should live in a focused value file.")
+        XCTAssertTrue(worktreeDraftsText.contains("struct QuillCodeWorktreePruneDraft"), "Worktree prune draft state should live in a focused value file.")
+        XCTAssertTrue(worktreeChromeText.contains("struct QuillCodeWorktreeChoiceSection"), "Shared worktree choice rows should live in focused worktree dialog chrome.")
+        XCTAssertTrue(worktreeChromeText.contains("struct QuillCodeWorktreeDialogFrame"), "Shared worktree sheet frame should live in focused worktree dialog chrome.")
+        XCTAssertTrue(worktreeChromeText.contains("QuillCodePressableButtonStyle"), "Worktree choice rows should use shared 0.96 press feedback.")
+        XCTAssertTrue(worktreeChromeText.contains("QuillCodeMetrics.minimumHitTarget"), "Worktree choice rows should preserve minimum hit targets.")
+        XCTAssertFalse(worktreeDialogsText.contains("struct QuillCodeWorktreeCreateDraft"), "Worktree dialogs should not own draft/request state.")
+        XCTAssertFalse(worktreeDialogsText.contains("struct QuillCodeWorktreeChoiceSection"), "Worktree dialogs should not own shared choice-row chrome.")
+        XCTAssertFalse(worktreeDialogsText.contains("struct QuillCodeWorktreeDialogFrame"), "Worktree dialogs should not own shared sheet chrome.")
         XCTAssertTrue(dialogChromeText.contains("struct QuillCodeDialogHeader"), "Shared dialog chrome should live in one reusable file.")
         XCTAssertTrue(renameDialogsText.contains("struct QuillCodeThreadRenameView"), "Rename sheets should remain in the small workspace rename dialog file.")
         XCTAssertFalse(renameDialogsText.contains("struct QuillCodeCommandPaletteView"), "Workspace rename dialogs should not own command palette UI.")
